+++ Application/Usecases/RaceExecution/GetRaceStandings.cs
@@ -13,7 +13,7 @@ public sealed record RaceStandingDto(
     int? GateNumber,
     string HorseName,
     string JockeyName,
-    int TotalPoints,
+    decimal TotalPoints,
     int LegWins,
     int LegTop3,
     int? Position);
